+++ inc/commoninjectionlib.class.php
@@ -2175,6 +2175,13 @@ class PluginDatainjectionCommonInjectionLib
                         $type_searchOptions[$id]['injectable'] = self::FIELD_INJECTABLE;
                     }
 
+                    //Some injection.class files are missing dropdown options. Set displaytype as dropdown if datatype is dropdown
+                    //$tmp['displaytype'] is still empty. Set to prevent overwriting on next IF
+                    if ((isset($tmp['datatype']) && $tmp['datatype'] == 'dropdown') && !isset($tmp['displaytype'])) {
+                        $type_searchOptions[$id]['displaytype'] = 'dropdown';
+                        $tmp['displaytype'] = 'dropdown';
+                    }
+
                     if (isset($tmp['linkfield']) && !isset($tmp['displaytype'])) {
                         $type_searchOptions[$id]['displaytype'] = 'text';
                     }