+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix default entity insertion for a user
 - Fixed `SQL` error when creating new injection model
+- Fixed issue with missing dropdown options
 
 ## [2.14.0] - 2024-10-10